Added type name and use it as discriminator for operators (+ string, int, double, - int, double).

+++ src/context.cpp
@@ -94,7 +94,7 @@ void Context::feed(std::string data){
 	auto spacepos=data.find_first_of(' ');
 	auto key=data.substr(0, spacepos);
 	auto value=data.substr(spacepos+1);
-	get_value(key).set(to_any(to_number(value)), *this);
+	get_value(key).set(to_any(int64_t(to_number(value))), *this);
 }
